+++ packages/parser/src/index.ts
@@ -3,3 +3,8 @@ export {
   parseSync as parseSync,
   loadModule as loadModule
 } from 'libpg-query';
+
+export {
+  deparse as deparse,
+  deparseSync as deparseSync,
+} from 'pgsql-deparser';