+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [17.7.4](https://github.com/launchql/pgsql-parser/compare/@pgsql/utils@17.7.3...@pgsql/utils@17.7.4) (2025-06-27)
+
+**Note:** Version bump only for package @pgsql/utils
+
+
+
+
+
+## [17.7.3](https://github.com/launchql/pgsql-parser/compare/@pgsql/utils@17.7.2...@pgsql/utils@17.7.3) (2025-06-26)
+
+**Note:** Version bump only for package @pgsql/utils
+
+
+
+
+
 ## [17.7.2](https://github.com/launchql/pgsql-parser/compare/@pgsql/utils@17.7.1...@pgsql/utils@17.7.2) (2025-06-24)
 
 **Note:** Version bump only for package @pgsql/utils